Move compiler inside backend
