--- src/qibolab/compilers/__init__.py
@@ -1,1 +1,0 @@
-from qibolab.compilers.rules import compiler
